+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to Dexter will be documented in this file.
 - Include '/' helper function for proxy URLs
 - Add export for SplitSwapRequest
 - Add tests for DexTransaction events
+- Fix `withSwapOutAmountMappings` for split swap requests
 
 ## [v4.1.0]
 - Support for multi-dex swap requests.